Move CBOR code from signedexchange to internal
