+++ go/bundle/encoder.go
@@ -13,7 +13,7 @@ import (
 	"strings"
 
 	"github.com/WICG/webpackage/go/bundle/version"
-	"github.com/WICG/webpackage/go/signedexchange/cbor"
+	"github.com/WICG/webpackage/go/internal/cbor"
 	"github.com/WICG/webpackage/go/signedexchange/structuredheader"
 )
 
@@ -178,7 +178,7 @@ func (is *indexSection) Finalize(ver version.Version) error {
 
 		mes := []*cbor.MapEntryEncoder{}
 		for url, es := range m {
-			if (len(es) > 1) {
+			if len(es) > 1 {
 				return errors.New("This WebBundle version '" + string(ver) + "' does not support variants, so we cannot have multiple resources per URL.")
 			}
 			me := cbor.GenerateMapEntry(func(keyE *cbor.Encoder, valueE *cbor.Encoder) {
