+++ go/internal/cbor/encoder.go
